Added httpsender script to handle anti CSRF token
At the moment ZAP supports only FORM parameter tokens. This script is designed to look for a CSRF token value in the response body using a regex pattern match. Besides capturing the CSRF token value from response body, it also uses the latest value for all outgoing requests irrespective of where the CSRF token is located i.e. FORM parameters or URL parameters or Cookie parameters.

This script is generic, however, it requires enough configuration setup for application under test. Following are several configuration entries that need to be modified in the script.
antiCsrfTokenName - Anti CSRF Token name in the application under test.
urlRegEx - Regex for the request URL that returns CSRF token value in its response body.
csrfTokenValueRegEx - Regex for capturing the CSRF token value from the response body. Make sure to use grouping in the regex, to capture a certain portion of the match.
matcherGroupNumber - Group number to be matched and captured as new Anti CSRF token value. This value will be replaced in all subsequent requests that has a CSRF token in the request.
parameterTypesList - An array of one or more parameter types to be looked in, for replacing the anti CSRF token value in the outgoing requests. Supported values in the array: formParamType, urlParamType, cookieParamType.