Add script to authenticate against MediaWiki sites using its API
MediaWiki protects against Login CSRF using a login token generated
on viewing the login page and storing it in the session and in a
hidden field in the login form. As a result, ZAP can't currently
handle MediaWiki logins with its form-based authentication. The
MediaWiki API can also be used to authenticate ZAP but it likewise
protects against login CSRF by returning a token on the first login
request which must be submitted on a second login request.