Verify scripts can be parsed
Add test to verify that the JavaScript and Jython scripts can be parsed
without errors.
Add Gradle build file (and wrapper) to run the tests, also, format Java
code and normalise the line endings of the scripts.
Remove Eclipse specific file and ignore it along with other IDE files,
those don't need to be tracked, they are automatically created/managed
by the corresponding IDE when the project is imported.
Add Travis CI build configuration file to automatically check the
scripts (under Java 8 and 11).