Prevent and warn on curl file inclusion
Use `--data-raw` instead of `data-binary` which does not allow local
file inclusion in the body.
Warn when file inclusion is being done in a custom header and do not
add the generated command to the clipboard (there does not seem to be
an alternative argument that does not use the `@file` syntax).

Thanks to James Kettle (@albinowax) for reporting.

Signed-off-by: thc202 <thc202@gmail.com>